Large cleanup of the code.
In particular, we make it much clearer which filename
is being passed around.  For the walker, we always make
it so that the filename passed into the callbacks is the
real filename; it is up to the callers to figure out
the got filename.  We also rename all of the variables
so that it is clear which filename is being passed.
We also move all command-line handling into the
individual commands to allow per-command checking.

Signed-off-by: Chris Lalancette <clalancette@irobot.com>